+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.5.0 - 2021-04-xx
+* Adopt breaking changes in Mockito 3.4.0 and higher (#4)
+* Introduced `KMockitoAnnotations.openMocks( Any )`
+* Deprecated `KMockitoAnnotations.initocks( Any )`
+
 ## [0.4.6] - 2021-03-31
 * Adjustment to package renaming of dependent mockito-kotlin library (#8)